[PERF] Add a parallel local CSV reader (#3055)
Adds a parallel CSV reader to speed up ingestion of CSV. This CSV reader
is used for local and uncompressed CSV files.

The approach in this PR adapts some ideas laid out in [1], namely the
idea of adjusting chunks of the CSV file with its neighbouring chunks so
that each chunk contains whole CSV records that can be decoded in
parallel.

However, as it turns out, the majority of performance gains came from
the use of buffer pools to minimize memory allocations.

## Some performance numbers

We consider a simple case of reading and performing `.collect()` on a
CSV file with 10^8 rows of 9 fields: 3 string fields, 5 int64 fields,
and 1 double field. This file is roughly 5GB in size.

```
Non-native executor:                 38.71140212500177s
Non-native executor, new CSV reader: 7.432862582994858s
Native executor:                     44.55550079200475s
Native executor, new CSV reader:     3.639032250008313s
```

This represents a roughly 12x speedup on CSV reads for the native
executor.

## Followups

Unfortunately, with this new CSV reader, the new native csv reader no
longer shows a small and stable amount of memory usage during an
aggregation. Memray shows a steady increase in resident set size that
disappears once the aggregation completes. For what it's worth, parsing
the whole CSV file using this reader then dumping the results does not
result in any memory increase. So it's possible we're simply not passing
results to downstream consumers in a way that they expect.

[1]: Ge, Chang et al. “Speculative Distributed CSV Data Parsing for Big
Data Analytics.” Proceedings of the 2019 International Conference on
Management of Data (2019).

---------

Co-authored-by: Sammy Sidhu <samster25@users.noreply.github.com>