[CHORE] Swordfish specific test fixtures (#3164)
This PR sets up a few swordfish related test fixtures, specifically:

- Parameterize `default_morsel_size = [1, None]` for dataframe tests
that do any into/repartitioning. This is to make sure that the operator
parallelism is working.
- Setup iteration tests in `test_iter.py`
- Makes the ordering assertions stricter on some tests. E.g. some tests
do `assert df.sort(col) == expected`, but there are other columns in df
that may not be sorted, and this won't be enough if morsel_size = 1.
This isn't a problem with swordfish but the test, where the sort should
actually involve more columns.

Big note: There was a problem with pivot not getting applied correctly.
This is because a dataframe pivot operation comprises of an agg + the
actual pivoting, but previously the pivot was implemented as an
intermediate operator, and the results of the agg were getting buffered.
In order for the pivot to work it has to receive all values with the
same group_by keys. This PR implements simplifies Pivot as a
BlockingSink, so all the work is in there.

---------

Co-authored-by: Colin Ho <colinho@Colins-MacBook-Pro.local>