[FEAT] Overwrite mode for write parquet/csv (#3108)
Addresses: https://github.com/Eventual-Inc/Daft/issues/3112 and
https://github.com/Eventual-Inc/Daft/issues/1768

Implements overwrite mode for write_parquet and write_csv.

Upon finishing the write, we are left with a manifest of written file
paths. We can use this to perform a `delete all files not in manifest`,
by:
1. Do an `ls` to figure out all the current files in the root dir.
2. Use daft's built in `is_in` expression to get the file paths to
delete.
3. Delete them.

Notes:
- Relies on fsspec for `ls` and `rm` functionalities. This is favored
over pyarrow filesystem because `rm` is a **bulk** delete method, aka we
can do the delete in a single API call. Pyarrow filesystem does not have
bulk deletes.

---------

Co-authored-by: Colin Ho <colinho@Colins-MacBook-Pro.local>
Co-authored-by: Colin Ho <colinho@Colins-MBP.localdomain>