[FEAT] Native Runner (#3178)
Makes swordfish a top level runner, `set_runner_native()`. Additionally
sets swordfish to be the default runner for development.

This PR also contains some bug fixes and test changes, of which I have
left comments for.

Additionally, this PR refactors swordfish in two ways:
1. Buffers scan tasks based on a `num_parallel_tasks` parameter, that
takes into account any pushed down limits.
2. Adds an `is_err` check on the sender in parts of the code where we
have a `while receiver.recv.await -> sender.send` pattern, such that it
breaks out of the loop if the sender is dropped. This is needed in cases
when the consumer is done receiving data, such as in a Limit, or if the
user is doing `iter(df)` and breaks out of the iter, which will cause
receivers to be dropped. As such, the senders should recognize this and
drop as well.

---------

Co-authored-by: Colin Ho <colinho@Colins-MacBook-Pro.local>
Co-authored-by: Colin Ho <colinho@Colins-MBP.localdomain>