[CHORE] Expose read_sql partition bound strategy and default to min-max (#3246)
Currently, read_sql calculates partition bounds using the
`PERCENTILE_DISC` function. However, this function does not scale well
to large tables, as it is an expensive window + sort function. A better
alternative is to take samples, then estimate partition bounds, as
described in this issue:
https://github.com/Eventual-Inc/Daft/issues/3245.

In the meantime, we should default to using the min-max calculations
instead, which was previously the fallback option.

---------

Co-authored-by: Colin Ho <colinho@Colins-MacBook-Pro.local>
Co-authored-by: Colin Ho <colinho@Colins-MBP.localdomain>