[BUG] Cleanup context side-effects (#3270)
Cleans up public-facing APIs on our context object that has
side-effects.

We should be very explicit in our context code when making state
changes. Eventually I think I want to deprecate the usage of
`daft.context.*` from users, and instead expose a `daft.connect(...)`
which will prevent user footguns.

Tests added in: https://github.com/Eventual-Inc/Daft/pull/3275

## Changes

1. In tests, use `get_tests_daft_runner_name` for checking the runner
instead of relying on the context
2. Refactors our context to be VERY explicit about when the runner is
created (`get_context().get_or_create_runner`)
3. I then refactored everywhere that was asking about state in the
context to first call `get_or_create_runner`, then access information
about the state.

Note that this might be a behavior change in some cases, but I think
that this is for the better. I.e. in order to access information about
the runner (mostly about what type of runner it is), we force clients to
materialize the runner.

I think in the future we should (after more testing) also refactor to:

- [ ] Let's remove the concept of RunnerConfig. When
`daft.context.set_runner_*` is called, we can just straight up set the
runner. Not sure why we need to lazily initialize the runner here.
- [ ] We can force initialization of the Runner at various key
entrypoints, namely `DataFrame.__init__`, `Expression.__init__` etc.
This is in effect saying -- when you use any Daft APIs, we are going to
capture the state of runner configurations and setting the runner.

---------

Co-authored-by: Jay Chia <jaychia94@gmail.com@users.noreply.github.com>