[FEAT] Monotonically Increasing Id for Swordfish (#3180)
Implements monotonically increasing id as a streaming sink with
`max_concurrency = 1`.

I tested multithreaded and single threaded implementations and found
that there was no performance gain in multithreaded. This is because
monotonically increasing id is a memory bound operator, all it does is
allocate an array of ints for the id. Multiple threads trying to do this
in parallel are bottlenecked by memory bandwidth.

It is actually also much simpler to implement this as a single threaded
operation, as we just need to keep a running count of the lengths of
morsels seen so far. This is effectively just `row_number`.

Note:
- There was a bug in the `pyfunc_into_table_iter` function, which
consumes python iterators in scan tasks (used in read_lance and
read_generator), where the consumer only calls `next()` on the iterator
once. This PR fixes that.

Co-authored-by: Colin Ho <colinho@Colins-MacBook-Pro.local>