[FEAT] Add ability to download log files from ray-cluster (#3406)
# Overview
- when submitting the job to the ray cluster, the extra
`DAFT_ENABLE_RAY_TRACING` env-var is set to `"1"`
- this will produce custom daft logs
- these log files are rsync'd down from the ray cluster, packaged up,
and sent back to the GHA Summary Page (from where they can be downloaded
by the user)

## Notes
When running the job on the ray-cluster, daft logs are *not* being
generated. Only ray logs are being generated. This is something that
will need to be investigated and resolved later.

For now, will request to have it merged, since everything else still
works (we can still grab ray logs, which is quite useful in and of
itself).