[CHORE] Remove CountMode and ResourceRequest from public API (#3429)
CountMode is used in
[`Expression.count`](https://www.getdaft.io/projects/docs/en/latest/api_docs/doc_gen/expression_methods/daft.Expression.count.html),
but it's not user friendly to have users figure out how to use a
CountMode object. Instead we get public methods to accept countmode
strings instead.

ResourceRequest is used in
[DataFrame.with_column](https://www.getdaft.io/projects/docs/en/latest/api_docs/doc_gen/dataframe_methods/daft.DataFrame.with_column.html)
and
[DataFrame.with_columns](https://www.getdaft.io/projects/docs/en/latest/api_docs/doc_gen/dataframe_methods/daft.DataFrame.with_columns.html)
but these are deprecated from public use for the same reasons as
CountMode. We thus remove ResourceRequest from the dataframe methods.