[FEAT] extend `build-commit` workflow to support different compile-archs (#3459)
# Overview
This PR extends the build commit workflow to also support building for
ARM architectures. The current workflow only built for x86.

## Limitations
- We are building using the `buildjet` service.
- Buildjet provides ARM machines. However, their ARM machines only run
Ubuntu 22.04.
  - Ubuntu 22.04 runs glibc2.34.
  - glibc is forwards-compatible, but *not* backwards compatible.
- This means that whenever you are running the built wheel, you must be
running it on a machine which contains glibc2.34 or greater. Installs of
the wheel will fail on platforms which contain older versions of glibc.
(Note that most AWS AMIs contain older versions of glibc).

## Notes
The storage structure of the wheels in AWS S3 is as below:
```txt
github-actions-artifacts-bucket
|- builds
   |- ${{ GITHUB.SHA }}
      |- ${{ WHEEL_NAME_FOR_X86 }}
      |- ${{ WHEEL_NAME_FOR_ARM }}
```