[FEAT]: expr simplifier (#3393)
still todo: 

- [x] add a bunch of test cases. 
~~- [ ] add simplification for `AND` exprs~~

### Note for reviewers:  

This bypasses the expr simplifier rule if the source is
`SQLScanOperator` it was simplifying the queries in a way that weren't
working with connectorx. I haven't been able to debug exactly what was
going on, but I think it's pretty ok to just skip this optimization for
this source for now.

Also, wanted to implement a const expr evaluator, but since our
`eval_expressions` is associated with `daft_table::Table`, this kind of
optimization is not currently possible, and some refactoring is needed.