refactor: create a rust based `PartitionSet`  (#3515)
most of this is ported from the python impl inside
`daft/runners/partitioning.py`.


### Note for reviewer. 

For context around why this is needed. The `DataFrame` class uses
`PartitionSet` extensively for various common operations such as `show`,
and `collect`. In order to add this functionality to our spark connect
implementation, we need a similar construct in rust.

Ideally, I'd like to port over the python implementation to use this new
rust one, but there are still a few things that I'm not entirely sure
how to implement (such as `RayPartitionSet`)

Not all of the methods inside `partitioning.rs` are used yet, But I
intend to follow up this PR with an implementation for
https://github.com/Eventual-Inc/Daft/issues/3498, and this is a
prerequisite as `show` relies on `get_preview_micropartitions`.