fix(parquet): Fix parquet reads of required fields nested within optional fields (#3598)
In https://dist-data.slack.com/archives/C041NA2RBFD/p1734108409644689,
it was discovered that when a parquet schema contained required fields
nested within optional fields, e.g.
```
  optional group struct_field = 10 {
    required binary nested_field (STRING) = 38;
  }
```
we would hit the following error:
```
validity mask length must match the number of values
```

Nested parquet fields can be marked as `required` within `optional`
parent fields because the parent field may be optional, but if the
parent field is defined then the nested field _must also be defined_.
`required`/`optional` affects the definition levels that encode parquet
values.

Internally, Daft also constructs `Required*` or `Optional*`[1] state
based on whether the field is marked as `required` vs `optional`
respectively. `Required*` state means that we don't rely on a validity
mask, whereas for `Optional*` state we maintain a validity mask.

However, when a parent field is `optional`, even if a nested field is
`required`, there can be situations where the nested value is NULL
because the parent value is NULL. In such situations, the `required`
nested field should keep track `Optional*` state (i.e. keep track of a
validity mask), but should still decode parquet definition levels as if
it were `required`.

In this PR, we keep track if any parent of a nested field is nullable.
In those cases we keep track of `Optional*` state.

[1]: E.g. `RequiredDictionary` vs `OptionalDictionary` states.