feat: Support intersect all and except distinct/all in DataFrame API (#3537)
This commits adds complete support for intersect and except operations.
To support that, it's consisted of:
1. Expose new methods such as `intersect_all`, `except_distinct` and
`except_all` in python side's DataFrame API
2. Add Except operator in set_operations and convert corresponding
logical plans
3. Complete intersect all branch with similar logic with except all
4. Add a new scalar function: `list_fill(num, val)` which will create a
list of num elements with the provided `val`

The intersect/except all logic is a bit complex and it will require
list_fill and explode to correctly compute the intersected or excepted
rows.