fix(swordfish): Track future poll times for explain analyze (#3511)
The explain analyzes and tracing spans don't properly track the
execution tasks. All the cpu times were showing up as 0.0, and the spans
weren't showing up in traces.

This is because the `execute`, `sink`, and `finalize` methods now return
futures, of which the timings are currently not tracked. This PR changes
it so that the time of the futures are now tracked.

---------

Co-authored-by: Colin Ho <colinho@Colins-MacBook-Pro.local>