chore: update PyO3 version to 0.23 (#3647)
Resolves #3641

The bulk of this PR are adding function signatures for python functions
defined in Rust. This is required now in 0.23 if the function has
optional parameters in order to distinguish it from nullable ones.

The other major change is that Python objects no longer implement
`Clone`, coming from PyO3 v0.22. This is because the reference counting
done to allow this actually can't happen properly without holding onto
the GIL. So instead, I wrapped pyobjects that need to be cloned with an
Arc. They can still be cloned with the GIL held using
`pyobj.clone_ref(py)`