fix(optimizer): Fix issues with join graph construction (#3668)
We make three fixes to join graph construction. With these changes, join
ordering using the naive left deep join orderer produces correct results
for TPCH queries.

### Push up projections and filters only if they sit in between joins

Instead of pushing up all Projections and Filters until we hit an
unreorderable node, we go down each linear chain of (reorderable)
Projections and Filters until we hit a Join node, then push up the
Projects and Filters we encountered along the way. If we hit an
unreorderable node first, we simply treat the operator that sits at the
top of the current linear chain as the relation to pass into the join.

For example, consider this query tree:
```
        //                InnerJoin (a = d)
        //                    /        \
        //                   /       Project
        //                  /        (d, quad <- double + double)
        //                 /               \
        //     InnerJoin (a = b)    InnerJoin (c = d)
        //        /        \           /         \
        // Scan(a)      Scan(b)   Filter        Scan(d)
        //                        (c < 5)
        //                           |
        //                        Project
        //                        (c <- c_prime, double <- c_prime + c_prime)
        //                           |
        //                        Filter
        //                        (c_prime > 0)
        //                           |
        //                        Scan(c_prime)
```
In between `InnerJoin(c=d)` and `Scan(c_prime)` there are Filter and
Project nodes. Since there is no join below `InnerJoin(c=d)`, we take
the `Filter(c<5)` operator as the relation to pass into the join (as
opposed to using `Scan(c_prime)` and pushing up the Projects and Filters
above it).

### If a relation needs to rename one column, make sure the other
columns are also selected

Previously, a relation might need to have a column renamed. However when
we did this, we didn't select the other columns in the relation, causing
some columns to be dropped prematurely.

### Rename columns for relations even if they are not involved in join
conditions

Previously, if we encountered a projection (e.g. `a_prime <- a`), we
would apply this projection directly above the source relation if
`a_prime` was involved in a join, e.g. `Join(left_on="a_prime", ..)`.
Now, we uniformly apply these projections regardless if the projection
affects a join condition.