refactor(execution): NativeExecutor refactor (#3689)
Small refactor to `NativeExecutor` to make it a bit more flexible for
spark connect.


some changes made

- make the following configurable:
  - runtime
  - progress_bar_manager
  - enable_explain_analyze
- refactor the engine to take in the plan when performing `.run` not at
creation.
- this is useful as we can now reuse an engine instead of having to
create a new one for every plan.
- Note: we could probably follow up with a refactor the python code to
have a single "native" engine instead of creating it every time
  
The runtime configuration is especially important for spark connect as
previously we were using `spawn_blocking` when running a plan, but now
we can leverage the existing compute runtime instead of creating a new
one.