feat(core): make micropartition streamable over tables (#3709)
### Description

makes Micropartition streamable over `Table`, and makes `Table` a bit
cheaper to clone by wrapping the `columns` column in an `Arc`.


The driving force for these changes is that currently the logic to go
from micropartition to table does not work nicely withing a streaming
context as the `get_tables` method returns `Arc<Vec<Table>>`. So you
can't easily chain streaming methods when working with micropartitions
and tables.