perf: Track accumulated selectivity in logical plan to improve probe side decisions (#3734)
Today, the cardinality of a join is estimated by taking the max of the
cardinality of the two sides because we assume primary key + foreign key
joins (pk-fk joins).

This is fine in the case where there is no filtering on the left or
right side of the join. However, when there is filtering, we
overestimate the cardinality of the join.

What's happening is that filtering one side of a pk-fk join should
reduce the cardinality of the join by the same proportion. However, in
the case where the unfiltered side has a larger cardinality, we
currently take its cardinality as the cardinality of the join, meaning
that the cardinality of the join is not scaled down by the proportion of
the filtered side.

This PR fixes this by tracking the accumulated selectivity of all
logical operations. When it comes time to estimate the cardinality of a
join, we scale down the estimated cardinality of the join by the
accumulated selectivity of the unchosen side (we don't need to scale
down by the selectivity of the chosen side because the cardinality has
already been scaled down prior to the join).

For a more concrete example, consider the following query tree:

```
                       InnerJoin
                     /           \
Filter(selectivity = 0.5)    Filter(selectivity = 0.9)
                   /                \
Scan(cardinality = 10)          Scan(cardinality = 10000)
```

For this inner join, the left side has a cardinality of 5 (10 * 0.5) and
the right side has a cardinality of 9000 (10000 * 0.9). Assuming a pk-fk
join, and ignoring filters, we expect the cardinality of the join to be
10000 (taking the max of the two sides to be the foreign key side).

However, if we take the filters into account, 50% of the primary keys
are filtered out and 90% of the foreign keys are filtered out. Assuming
that the filters are independent, we expect the cardinality of the join
to be 10000 * 0.5 * 0.9 = 4500. To make this estimate, we take the
larger cardinality of the two sides (9000) and scale it down by the
selectivity of the smaller side (0.5) which gives us the expected
cardinality of the join (4500).

This is an approximation, and it might well be the case that the
selectivities are not independent. This could be handled in the future
by using histograms, or we could also penalize multiplication of the
selectivity by some tuned penalty.

## Results with TPCH

These changes speed up TPC-H queries 8, 9, and 10, which were suffering
from incorrect probe side decisions.

Q5 saw some regressions because the cardinalities for two sides were
similar enough that these changes caused the probe side to flip
incorrectly. However, we currently have an incorrect join order for Q5,
so once join reordering is turned on this may no longer be an issue.

Other queries were not affected.