refactor: port DaftContext to rust side (#3767)
ports the `DaftContext` to the rust side so we can more easily reuse it
for connect & catalog work.


notes for reviewers: 
- after removing `daft-local-execution` from daft-connect, it made
`cargo check` fail, so some of the changes to unrelated `Cargo.toml`'s
was to properly feature flag them. We missed a couple that were causing
things to break.
- @rchowell I requested review from you specifically to take a look at
[context.rs](https://github.com/Eventual-Inc/Daft/pull/3767/files#diff-5b536482a8303505c0c91a561a632218591bf8f02fff21f9f8a3535a3f0ff8a5).
This contains the "context"/"session", and the global state that was
previously only available in python. I added the `catalog` to this state
as well.