perf: Use parquet metadata from schema inference for accurate scan task statistics (#3784)
When performing a glob scan on a parquet file, if schema inference is
enabled, we read the schema from the parquet file's metadata. This
metadata also comes with cardinality information. We can use this
cardinality information to populate the first scan task's cardinality
accurately.

If the scan operator produces a single scan task, then this allows us to
have perfectly correct cardinality information (minus pushdowns) for the
given scan node.