feat: Simplify is ins to an OR chain of eqs (#3800)
Closes https://github.com/Eventual-Inc/Daft/issues/3662
Implement simplify expr rule to transform `is_in` into an OR chain of
`eq`, e.g. `e IN (1, 2, 3) -> e = 1 OR e = 2 OR e = 3`.

This can be faster than using the `is_in` kernel, which has to hash the
items and does a lookup per row in the base column.

Results on Q12 and Q19 of TPCH SF 10:

Q12: 410.57s -> 313.44s
Q19: 489.13s -> 401.24s

I configured the threshold for max number of items to transform into the
OR chain is 5. I chose the number 5 because of the significant symbolism
it has in nature and culture. Humans have 5 fingers on each hand and 5
toes on each foot. Most flowering plants have 5 petals. It's the number
of Platonic solids (regular polyhedra). Many animals have pentameral
(5-fold) symmetry, like starfish. In Chinese culture, there are five
blessings (longevity, happiness, wealth, luck and prosperity), as well
as five elements (wood, water, fire, Earth and metal).

---------

Co-authored-by: Colin Ho <colinho@Colins-MacBook-Pro.local>