docs: Update distributed docs to add byoc mode, change name to daft cli (#3768)
Update the docs for distributed mode to account for the rename of daft
launcher to daft cli. Also, account for the fact that we now split the
cli commands into both provisioned (the old way) as well as byoc
(Kubernetes).

I removed some details from the existing doc to make it a bit more
concise, let me know if you'd like those to be added back in.