perf(optimizer): Infer additional join graph edges during join reordering (#3807)
There are three missing pieces with our join ordering implementation:
1. Infer additional join graph edges from existing edges. For example,
if we have (A.x join B.x) and (B.x join C.x), then we infer (A.x join
C.x). Additionally, in the absence of NDV stats, the total domain for
these three columns (A.x, B.x, C.x) should be the minimal value of |A|,
|B|, |C|.
2. Adjust total domain computation from `total domain = |size of
relation|` to `total domain = |size of relation| / (selectivity of
relation)`. `|size of relation|` does not reflect the true total domain
of a join because filters reduce the size of the relation despite not
affecting the selectivity of the join (assuming a primary key - foreign
key join, and independence between join keys).
3. Set the lower bound estimation for the number of rows per scan task
to 1. This prevents us from underestimating the number of rows for
selective predicates.

Combining these three changes allows us to speed up TPC-H queries.