refactor!: split column expression into unresolved and resolved types (#3804)
BREAKING CHANGE: Python struct get syntactic sugar is no longer allowed
(ex: `col("a.b")` -> `col("a").struct.get("b")`.

I did not think it was worth continuing to maintain what is essentially
a small compiler for an undocumented and somewhat unclear feature. I
would have to make some significant code changes to get it to work with
unresolved columns, and I figured I would just rip it out. In the future
I'd like to support something like this instead: `col("a")["b"]`.

---

This PR is the first step to non-equality joins in Daft. We introduce an
`UnresolvedColumn` type so that we can resolve it in the logical plan
builder into a `ResolvedColumn::Basic`, `ResolvedColumn::JoinSide` (for
joins), or `ResolvedColumn::OuterRef` (for subqueries). In addition, we
introduce a `SubqueryAlias` logical op that is used to determine the
sources for `UnresolvedColumns` with `plan_ref` set.

Along with this PR is a refactor of our SQL planner. The `SubqueryAlias`
op is now used to encode naming and scoping information, so the
`Relation` struct is no longer necessary.

Additionally, I discovered that we can significantly simplify
`SQLPlanner::plan_non_agg_query` and `SQLPlanner::plan_aggregate_query`.
I believe the prior complexity came due to the fact that in many clauses
(GROUP BY, ORDER BY, etc), you can reference both columns in the input
tables as well as aliases in the select clause. However, the planner's
`bound_columns` already solves this issue.

This may yield slightly less performant plans due to redundant
expressions, but that should be handled by the optimizer instead of the
SQL planner. Moreover, I was able to remove a sort from
`plan_aggregate_query`, which may yield both performance and correctness
improvements, since the prior code assumed you can split a sort up into
two, however that may lead to different results due to sort orders
(`ORDER BY a, b` != `ORDER BY b, a`) as well as whether our sort impl is
stable or unstable.