feat: Enable capturing and sending logs (#3875)
# Overview

This PR enables capturing logs when running on the native runner and
broadcasting it over HTTP to port 3238.

Now, all printouts directed towards `STDOUT` will be captured and stored
in some internal buffer. When the HTTP broadcast message is emitted, all
of the captured logs for *that query* will be sent alongside the rest of
the query information.

## Usage

Create a python file with the following contents:

```py
# my_daft_script.py

import daft

print("Hello, world!")
daft.from_pydict({}).collect()
```

Start your dashboard and run the python script:

```sh
daft-dashboard launch
# don't forget to run `daft-dashboard shutdown` at the end!

DAFT_DASHBOARD=1 python my_daft_script.py
```

The logs should be captured. You can test that this is true (without
using the UI) by running:

```sh
curl http://localhost:3238/api/queries
```

This should return to you a list with one object in it. That object
should contain the key-pair `"logs": "Hello, world!\n"` inside of it.

## Notes

Although the functionality on the backend is implemented, the UI will
still not display any logs just yet. *That* functionality will be coming
in a separate PR. However, this PR will *not* break *any* existing
functionality! This is because the new key-pair will just be ignored by
the frontend.

One of the nice things of full-stack development is that sometimes
changes to the backend/frontend can be made incrementally and merged in
without affecting the other.