perf: Fix join cost estimates (#3831)
This PR fixes join cost estimates in the following ways:
- Uses number of rows instead of size in bytes for making probe side
decisions
- Decreases the aggressiveness of cardinality reductions when there is
an `is not NULL` filter
- Decreases the aggressiveness of inequality filters, especially since
these often appear as pairs of a "between" clause
- Increases the aggressiveness of exact equality filters.
- For join edges where there are more than one join condition, adjust
cost estimations so that we compute the total domain as `|(key1, key2,
...)| ~= min(|left side|, |right side|)`.
With multiple join conditions, we know that the join is not a pk-fk join
on each join key. Rather, it's a pk-fk join on the tuple of join keys,
which we estimate as the cardinality of the smaller table of the join.